+++ packages/ai-provider/src/hf-transformers/registry/HFT_Inline_RegisterJobFns.ts
@@ -6,6 +6,9 @@
 //    *******************************************************************************
 
 import { getAiProviderRegistry } from "@podley/ai";
+import { env } from "@sroussey/transformers";
+
+import { HF_TRANSFORMERS_ONNX } from "../common/HFT_Constants";
 import {
   HFT_Download,
   HFT_TextEmbedding,
@@ -15,8 +18,6 @@ import {
   HFT_TextSummary,
   HFT_TextTranslation,
 } from "../common/HFT_JobRunFns";
-import { HF_TRANSFORMERS_ONNX } from "../common/HFT_Constants";
-import { env } from "@sroussey/transformers";
 
 export async function register_HFT_InlineJobFns() {
   // @ts-ignore