+++ packages/task-graph/src/common.ts
@@ -5,22 +5,23 @@
 //    *   Licensed under the Apache License, Version 2.0 (the "License");           *
 //    *******************************************************************************
 
-export * from "./task/TaskTypes";
-export * from "./task/TaskSchema";
-export * from "./task/TaskError";
-export * from "./task/ITask";
-export * from "./task/TaskEvents";
-export * from "./task/TaskJSON";
-export * from "./task/Task";
+export * from "./task/ArrayTask";
 export * from "./task/GraphAsTask";
 export * from "./task/GraphAsTaskRunner";
-export * from "./task/TaskRegistry";
+export * from "./task/ITask";
+export * from "./task/JobQueueFactory";
 export * from "./task/JobQueueTask";
+export * from "./task/Task";
+export * from "./task/TaskError";
+export * from "./task/TaskEvents";
+export * from "./task/TaskJSON";
 export * from "./task/TaskQueueRegistry";
-export * from "./task/ArrayTask";
+export * from "./task/TaskRegistry";
+export * from "./task/TaskSchema";
+export * from "./task/TaskTypes";
 
-export * from "./task-graph/DataflowEvents";
 export * from "./task-graph/Dataflow";
+export * from "./task-graph/DataflowEvents";
 
 export * from "./task-graph/ITaskGraph";
 export * from "./task-graph/TaskGraph";
@@ -28,10 +29,10 @@ export * from "./task-graph/TaskGraphEvents";
 export * from "./task-graph/TaskGraphRunner";
 
 export * from "./task-graph/Conversions";
-export * from "./task-graph/Workflow";
 export * from "./task-graph/IWorkflow";
+export * from "./task-graph/Workflow";
 
-export * from "./storage/TaskOutputRepository";
-export * from "./storage/TaskOutputTabularRepository";
 export * from "./storage/TaskGraphRepository";
 export * from "./storage/TaskGraphTabularRepository";
+export * from "./storage/TaskOutputRepository";
+export * from "./storage/TaskOutputTabularRepository";