+++ packages/test/src/samples/index.ts
@@ -5,22 +5,20 @@
 //    *   Licensed under the Apache License, Version 2.0 (the "License");           *
 //    *******************************************************************************
 
-import { AiJob, AiProviderInput } from "@podley/ai";
+import { AiJob, AiJobInput } from "@podley/ai";
 import { TENSORFLOW_MEDIAPIPE } from "@podley/ai-provider";
 import { ConcurrencyLimiter, JobQueue } from "@podley/job-queue";
-import { getTaskQueueRegistry, TaskInput, TaskOutput } from "@podley/task-graph";
 import { InMemoryQueueStorage } from "@podley/storage";
+import { getTaskQueueRegistry, TaskInput, TaskOutput } from "@podley/task-graph";
 export * from "./MediaPipeModelSamples";
 export * from "./ONNXModelSamples";
 
 export async function register_HFT_InMemoryQueue() {
-  const jobQueue = new JobQueue<AiProviderInput<TaskInput>, TaskOutput>(
+  const jobQueue = new JobQueue<AiJobInput<TaskInput>, TaskOutput>(
     "HF_TRANSFORMERS_ONNX",
-    AiJob<TaskInput, TaskOutput>,
+    AiJob<AiJobInput<TaskInput>, TaskOutput>,
     {
-      storage: new InMemoryQueueStorage<AiProviderInput<TaskInput>, TaskOutput>(
-        "HF_TRANSFORMERS_ONNX"
-      ),
+      storage: new InMemoryQueueStorage<AiJobInput<TaskInput>, TaskOutput>("HF_TRANSFORMERS_ONNX"),
       limiter: new ConcurrencyLimiter(1, 10),
     }
   );
@@ -29,13 +27,11 @@ export async function register_HFT_InMemoryQueue() {
 }
 
 export async function register_TFMP_InMemoryQueue() {
-  const jobQueue = new JobQueue<AiProviderInput<TaskInput>, TaskOutput>(
+  const jobQueue = new JobQueue<AiJobInput<TaskInput>, TaskOutput>(
     TENSORFLOW_MEDIAPIPE,
-    AiJob<TaskInput, TaskOutput>,
+    AiJob<AiJobInput<TaskInput>, TaskOutput>,
     {
-      storage: new InMemoryQueueStorage<AiProviderInput<TaskInput>, TaskOutput>(
-        TENSORFLOW_MEDIAPIPE
-      ),
+      storage: new InMemoryQueueStorage<AiJobInput<TaskInput>, TaskOutput>(TENSORFLOW_MEDIAPIPE),
       limiter: new ConcurrencyLimiter(1, 10),
     }
   );