+++ packages/sqlite/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @podley/sqlite
 
+## 0.0.38
+
 ## 0.0.37
 
 ### Patch Changes