+++ packages/util/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @podley/util
 
+## 0.0.38
+
 ## 0.0.37
 
 ### Patch Changes